Line Graph Update (#44)
* Visual LG Changes

* Add graphql to Linegraph

* lg changes

* Minor changes to LG

* minor fix

+++ mock/graphql/schema.js
@@ -4,7 +4,7 @@ type Descriptor {
 	value: Float!
 	longitude: Float!
 	latitude: Float!
-	unixTimestamp: Int!
+	unixTimestamp: Float!
 }
 
 type Query {
